backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.20
asyncpg==0.26.0
pydantic==2.6.0
python-dotenv==1.0.0
boto3==1.30.0
celery[redis]==5.4.0
redis==4.5.5
transformers==4.40.2
torch==2.1.0
pillow==10.0.0
prometheus-fastapi-instrumentator==6.0.2
sentry-sdk==1.29.0
